+++ apps/jest-expo-mock-generator/App.tsx
@@ -11,6 +11,7 @@ const blacklist = [
   'ExpoLinking',
   'ExpoFont',
   'ExpoFileSystem',
+  'ExpoVideo',
 ];
 
 type ModuleRegistrySchema = [